pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "organism-tractability"
version = "0.0.1"
description = "Publication snapshot: fetch tractability features for organisms (CSV in / CSV out)"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Arcadia Science", email = "swe@arcadiascience.com" }]

dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.11.9",
  "firecrawl-py>=4.3.6",
  "click>=8.3.0",
  "tenacity>=9.1.2",
  "pyyaml>=6.0.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
  "pre-commit>=4.2.0",
  "pyright>=1.1.401",
  "pytest>=8.3.5",
  "ruff>=0.11.11",
]

[tool.ruff]
src = ["."]
line-length = 100
indent-width = 4

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F403", "F405"]

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP", "W"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
order-by-type = true
no-lines-before = ["future", "standard-library"]

